+++ src/lib.rs
@@ -113,11 +113,11 @@ pub use ifo::{
     CellPlaybackInfo, CellPositionInfo, CellType, DvdChapter, DvdTitle, DvdTitleEntry, FrameRate,
     McExtensionEntry, MenuAttributes, MenuType, NavCommand, PaletteEntry, Pgc, PgcCommandTable,
     PgcTime, Pgci, PgciLu, PgciLuSrp, PgciSrp, PgciUt, PgciUtSrp, PlaybackMode, PtlMait, Ptt,
-    PttTitle, StillTime, SubpictureAttributes, SubpictureCodingMode, SubpictureLanguageType,
-    SubpictureStreamControl, TitleAttributes, TmapEntry, TtSrpt, VideoAspectRatio, VideoAttributes,
-    VideoCodingMode, VideoResolution, VideoStandard, VmgIfo, VmgPtlMait, VmgVtsAtrt,
-    VmgVtsAtrtEntry, VobuAdmap, VtsCAdt, VtsIfo, VtsPttSrpt, VtsTmap, VtsTmapti, VtsiMat,
-    DVD_SECTOR, VMG_MAGIC, VTS_MAGIC,
+    PttTitle, StillTime, SubpictureAttributes, SubpictureCodingMode, SubpictureDisplay,
+    SubpictureLanguageType, SubpictureStreamControl, TitleAttributes, TmapEntry, TtSrpt,
+    VideoAspectRatio, VideoAttributes, VideoCodingMode, VideoResolution, VideoStandard, VmgIfo,
+    VmgPtlMait, VmgVtsAtrt, VmgVtsAtrtEntry, VobuAdmap, VtsCAdt, VtsIfo, VtsPttSrpt, VtsTmap,
+    VtsTmapti, VtsiMat, DVD_SECTOR, VMG_MAGIC, VTS_MAGIC,
 };
 pub use iso9660::{
     DirectoryRecord, Iso9660Entry, Iso9660Volume, PathTableEntry, PrimaryVolumeDescriptor,